refactor: 業務ロジックの分割をするためのファイル分け
